+++ backend/dqt/settings/docker.py
@@ -37,3 +37,11 @@ RABBIT = {
     "password": os.getenv("RABBIT_PASSWORD"),
     "exchange_name": os.getenv("RABBIT_EXCHANGE_NAME"),
 }
+
+GDOCS_TEMPLATES = {
+    "DEFAULT_BASE_TEMPLATE" : os.getenv("DEFAULT_BASE_TEMPLATE"),
+    "DEFAULT_DATASET_TEMPLATE" :  os.getenv("DEFAULT_DATASET_TEMPLATE"),
+    "DEFAULT_FIELD_TEMPLATE":  os.getenv("DEFAULT_FIELD_TEMPLATE"),
+    "DEFAULT_RESOURCE_TEMPLATE":  os.getenv("DEFAULT_RESOURCE_TEMPLATE"),
+    "DEFAULT_ERROR_TEMPLATE" :  os.getenv("DEFAULT_ERROR_TEMPLATE"),
+}
