+++ backend/dqt/tools/tags/template_tags/dataset.py
@@ -1,6 +1,6 @@
 from django.utils.translation import gettext as _
 from dqt.models import DatasetLevelCheck
-from dqt.settings.base import GDOCS_TEMPLATES
+from dqt.settings.docker import GDOCS_TEMPLATES
 from dqt.tools.errors import CheckNotComputedError
 from dqt.tools.misc import terms_enumeration
 from dqt.tools.tags.leaf_tags.dataset.bar.count import CountLeafTag as bar_CountLeafTag
